uhd: remove integer conversions and carry floats throughout
The UHD driver will perform an internal float conversion
from the over-the-wire type, which is then output through
the API as integer, which is later converted in the radio
interface back to float. This patch removes some of the
type conversion madness at the expense of doubling I/O
bandwidth at the lower transceiver layer, but this
appears to to be worthwhile tradeoff when squeezing
out performance.

Obviously, this breaks a number of interfaces and is
incompatible with non-UHD builds.

+++ public-trunk/Transceiver/USRPping.cpp
@@ -57,24 +57,24 @@ int main(int argc, char *argv[]) {
   LOG(INFO) << "Looping...";
   bool underrun;
 
-  //short data[]={0x00,0x02};
+  //float data[]={0x00,0x02};
 
   usrp->updateAlignment(20000);
   usrp->updateAlignment(21000);
 
   int numpkts = 2;
-  short data2[156*2*numpkts];
+  float data2[156*2*numpkts];
   for (int i = 0; i < 156*numpkts; i++) {
     data2[i<<1] = 10000;//4096*cos(2*3.14159*(i % 126)/126);
     data2[(i<<1) + 1] = 10000;//4096*sin(2*3.14159*(i % 126)/126);
   }
 
   for (int i = 0; i < 1; i++) 
-    usrp->writeSamples((short*) data2,156*numpkts,&underrun,102000+i*1000);
+    usrp->writeSamples((float*) data2,156*numpkts,&underrun,102000+i*1000);
 
   timestamp = 19000;
   while (1) {
-    short readBuf[512*2];
+    float readBuf[512*2];
     int rd = usrp->readSamples(readBuf,512,&underrun,timestamp);
     if (rd) {
       LOG(INFO) << "rcvd. data@:" << timestamp;
